Split Azure Devops unit tests into multiple modules, consistent with the split of the Azure Devops collector into multiple modules.
